permssion-updated

+++ routes/web.php
@@ -22,31 +22,34 @@ Route::middleware('checkauth')->group(function () {
   Route::get('/admin/dashboard', [AdminController::class, 'dashboard'])->name('admin.dashboard');
   Route::post('/admin/logout', [AuthController::class, 'logout'])->name('admin.logout');
 
-  // Subuser
-  Route::get('/admin/index', [SubUserController::class, 'index'])->name('subuser.index');
-  Route::get('/admin/create', [SubUserController::class, 'create'])->name('subuser.create');
-  Route::post('/admin/store', [SubUserController::class, 'store'])->name('subuser.store');
-  Route::get('/admin/subuser/{subuser}/edit', [SubUserController::class, 'edit'])->name('subuser.edit');
-  Route::patch('/admin/subuser/{subuser}/update', [SubUserController::class, 'update'])->name('subuser.update');
-  Route::delete('/admin/subuser/{subuser}/delete', [SubUserController::class, 'destroy'])->name('subuser.delete');
-  Route::get('/subuser/status', [SubUserController::class, 'status'])->name('subuser.status');
+  // Subuser Route: Start
+  Route::get('/admin/index', [SubUserController::class, 'index'])->middleware('permission:viewAny,subusers')->name('subuser.index');
+  Route::get('/admin/create', [SubUserController::class, 'create'])->middleware('permission:create,subusers')->name('subuser.create');
+  Route::post('/admin/store', [SubUserController::class, 'store'])->middleware('permission:create,subusers')->name('subuser.store');
+  Route::get('/admin/subuser/{subuser}/edit', [SubUserController::class, 'edit'])->middleware('permission:update,subusers')->name('subuser.edit');
+  Route::patch('/admin/subuser/{subuser}/update', [SubUserController::class, 'update'])->middleware('permission:update,subusers')->name('subuser.update');
+  Route::delete('/admin/subuser/{subuser}/delete', [SubUserController::class, 'destroy'])->middleware('permission:delete,subusers')->name('subuser.delete');
+  Route::get('/subuser/status', [SubUserController::class, 'status'])->middleware('permission:update,subusers')->name('subuser.status');
+  // Subuser Route: End
 
-  // Role
-  Route::get('/admin/role/index', [RoleController::class, 'index'])->name('role.index');
-  Route::get('/admin/role/create', [RoleController::class, 'create'])->name('role.create');
-  Route::post('/admin/role/store', [RoleController::class, 'store'])->name('role.store');
-  Route::get('/admin/role/{role}/edit', [RoleController::class, 'edit'])->name('role.edit');
-  Route::patch('/admin/role/{role}/update', [RoleController::class, 'update'])->name('role.update');
-  Route::delete('/admin/role/{role}/delete', [RoleController::class, 'delete'])->name('role.delete');
+  // Role Route: Start
+  Route::get('/admin/role/index', [RoleController::class, 'index'])->middleware('permission:viewAny,roles')->name('role.index');
+  Route::get('/admin/role/create', [RoleController::class, 'create'])->middleware('permission:create,roles')->name('role.create');
+  Route::post('/admin/role/store', [RoleController::class, 'store'])->middleware('permission:create,roles')->name('role.store');
+  Route::get('/admin/role/{role}/edit', [RoleController::class, 'edit'])->middleware('permission:update,roles')->name('role.edit');
+  Route::patch('/admin/role/{role}/update', [RoleController::class, 'update'])->middleware('permission:update,roles')->name('role.update');
+  Route::delete('/admin/role/{role}/delete', [RoleController::class, 'delete'])->middleware('permission:delete,roles')->name('role.delete');
+  // Role Route: End
 
-  // Builder
-  Route::get('/admin/builder/{builder}', [BuilderController::class, 'show'])->whereNumber('builder')->name('builder.show');
-  Route::get('/admin/builder/index', [BuilderController::class, 'index'])->name('builder.index');
-  Route::get('/admin/builder/create', [BuilderController::class, 'create'])->name('builder.create');
-  Route::post('/admin/builder/store', [BuilderController::class, 'store'])->name('builder.store');
-  Route::get('/admin/builder/{builder}/edit', [BuilderController::class, 'edit'])->name('builder.edit');
-  Route::patch('/admin/builder/{builder}/update', [BuilderController::class, 'update'])->name('builder.update');
-  Route::delete('/admin/builder/{builder}/destroy', [BuilderController::class, 'destroy'])->name('builder.destroy');
-  Route::delete('/builder/bulk-delete', [BuilderController::class, 'bulkDelete'])
-    ->name('builder.bulkDelete');
+  // Builder Route: Start
+  Route::get('/admin/builder/{builder}/show', [BuilderController::class, 'show'])->middleware('permission:view,builder')->name('builder.show');
+  Route::get('/admin/builder/index', [BuilderController::class, 'index'])->middleware('permission:viewAny,builder')->name('builder.index');
+  Route::get('/admin/builder/create', [BuilderController::class, 'create'])->middleware('permission:create,builder')->name('builder.create');
+  Route::post('/admin/builder/store', [BuilderController::class, 'store'])->middleware('permission:create,builder')->name('builder.store');
+  Route::get('/admin/builder/{builder}/edit', [BuilderController::class, 'edit'])->middleware('permission:update,builder')->name('builder.edit');
+  Route::patch('/admin/builder/{builder}/update', [BuilderController::class, 'update'])->middleware('permission:update,builder')->name('builder.update');
+  Route::delete('/admin/builder/{builder}/destroy', [BuilderController::class, 'destroy'])->middleware('permission:delete,builder')->name('builder.destroy');
+  Route::delete('/builder/bulk-delete', [BuilderController::class, 'bulkDelete'])->middleware('permission:delete,builder')->name('builder.bulkDelete');
+  // Builder Route: End
+
 });
